Refactor dataset, move test stuff from dataset to dataset_test
