+++ 3d_cnn_prometheus/learning/config.json
@@ -6,7 +6,7 @@
     "images_dir": "images",
     "norm": true,
     "step": 3,
-    "window": [
+    "chunk_size": [
         16,
         32,
         32
