+++ 3d_cnn_prometheus/learning/model/experiment_setup.py
@@ -1,0 +1,4 @@
+from sacred import Experiment
+
+ex = Experiment()
+ex.add_config("/home/alikrz/Pulpit/MyStuff/cancer-detection/3d-cnn-prometheus/3d_cnn_prometheus/learning/config.json")
