+++ 3d_cnn_prometheus/learning/model/utils.py
@@ -1,14 +1,13 @@
 import os
+
 import numpy as np
 import tensorflow as tf
 import tensorflow.keras.backend as K
 import tensorflow_probability as tfp
-from sacred import Experiment
 from tensorflow.keras.callbacks import Callback
 from tensorflow.keras.losses import binary_crossentropy
 
-ex = Experiment()
-ex.add_config("/home/alikrz/Pulpit/MyStuff/cancer-detection/3d-cnn-prometheus/3d_cnn_prometheus/learning/config.json")
+from .experiment_setup import ex
 
 
 def round_down(num, factor):
@@ -54,7 +53,7 @@ def variational_free_energy_loss(model, scale_factor, kl_alpha):
     # KL Divergence should be applied once per epoch only, so
     # scale_factor should be num_samples / batch_size.
     kl = sum(model.losses) / scale_factor
-    
+
     def loss(y_true, y_pred):
         bce = binary_crossentropy(y_true, y_pred)
         return bce + K.get_value(kl_alpha) * kl
@@ -75,6 +74,7 @@ def normal_prior(prior_std):
 
     return prior_fn
 
+
 def get_latest_file(directory, match=""):
     """Gets the absolute file path of the last modified file in a directory.
 
@@ -87,6 +87,7 @@ def get_latest_file(directory, match=""):
         return max(paths, key=os.path.getctime)
     else:
         return None
+
 
 class AnnealingCallback(Callback):
     def __init__(self, kl_alpha, kl_start_epoch, kl_alpha_increase_per_epoch):
